+++ src/utils.js
@@ -296,3 +296,11 @@ export function parseErrorMessage(errorMessage) {
       </p>
     ));
 }
+
+export const getTlrSettings = (settings) => {
+  try {
+    return JSON.parse(settings);
+  } catch (error) {
+    return {};
+  }
+};